% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{index}
\alias{index}
\title{Aggregate Index}
\format{An object of class \code{data.frame} with 100 rows and 2 columns.}
\usage{
index
}
\description{
An aggregate index, in this case the Consumer Price Index, and its associated date.
}
\keyword{datasets}
