% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexr.R
\name{indexr}
\alias{indexr}
\title{Calculate an Aggregate Index}
\usage{
indexr(w0, sub0, index0, sub1)
}
\arguments{
\item{w0}{the reference set of weights.}

\item{sub0}{the reference set of subindices.}

\item{index0}{the reference value of the index.}

\item{sub1}{the current set of subindices.}
}
\value{
The aggregate index value, or vector of values, depending on the arguments supplied.
}
\description{
This function calculates an aggregate index, such as CPI, from its reference and current subindices, their reference and current relative importance weights, and a reference value of the aggregate index.
}
