% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subindices}
\alias{subindices}
\title{Example Subindices (ECs)}
\format{An object of class \code{data.frame} with 100 rows and 69 columns.}
\usage{
subindices
}
\description{
The subindices, and associated date, used to calculate aggregate index; in this case Expenditure Classes to Calculate CPI.
}
\keyword{datasets}
