% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightr.R
\name{weightr}
\alias{weightr}
\title{Calculate Relative Importance Weights}
\usage{
weightr(w0, sub0, index0, sub1, index1)
}
\arguments{
\item{w0}{the reference set of weights.}

\item{sub0}{the reference set of subindices.}

\item{index0}{the reference value of the index.}

\item{sub1}{the current set of subindices.}

\item{index1}{the currewnt value of the index}
}
\value{
A vector of weights
}
\description{
This function calculates the relative importance weights for a collection of subindices, based on changes in the subindices and the aggregate index.
}
