% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights}
\alias{weights}
\title{2009 Relative Importance Weights}
\format{An object of class \code{data.frame} with 1 rows and 69 columns.}
\usage{
weights
}
\description{
The December 2009 (first date of 'index' and 'subindices') official relative importance weights, by Expenditure Class (column names), used to calculate the aggregate index.
}
\keyword{datasets}
