% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexr.R
\name{indexr}
\alias{indexr}
\title{Calculate an Aggregate Index}
\usage{
indexr(w0, sub0, index0, sub1)
}
\arguments{
\item{w0}{the reference set of weights.}

\item{sub0}{the reference set of subindices.}

\item{index0}{the reference value of the index.}

\item{sub1}{the current set of subindices.}
}
\value{
The aggregate index value, or vector of values, depending on the arguments supplied (vector/scalar).
}
\description{
This function calculates an aggregate index, such as CPI, from (i) its reference and current subindices, (ii) its reference and current relative importance weights, and (iii) a reference value of the aggregate index.
}
\examples{
\dontrun{
# Calculate Month 2 (2010-01-01) index using:
(i) Month 2 subindices; and
(ii) Month 1 (2009-12-01) weights, subindices,
and aggregate index

indexr(
  w0 = weights[1,-1],
  sub0 = subindices[1,-1],
  index0 = index[1,2],
  sub1 = subindices[2,-1]
  )

# Make sure the date for reference weight/sub/index match;
i.e.:
weights[1,'Date'] == subindices[1,'date'] &
subindices[1,'date'] == index[1,'date']
}
}
